+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `nova-spotify-auth-tool` will be documented in this file.
 
+## 0.1.5 - 2018-11-26
+
+- Spotify Facade.
+
 ## 0.1.4 - 2018-11-26
 
 - Request Spotify user ID.